refactor: Collect results manipulations in a single module
